+++ packages/addons/lucia/index.ts
@@ -26,13 +26,13 @@ type Dialect = 'mysql' | 'postgresql' | 'sqlite' | 'turso';
 let drizzleDialect: Dialect;
 let schemaPath: string;
 
-const options = defineAddonOptions({
-	demo: {
+const options = defineAddonOptions()
+	.add('demo', {
 		type: 'boolean',
 		default: true,
 		question: `Do you want to include a demo? ${colors.dim('(includes a login/register page)')}`
-	}
-});
+	})
+	.build();
 
 export default defineAddon({
 	id: 'lucia',
@@ -48,8 +48,8 @@ export default defineAddon({
 	run: ({ sv, typescript, options, kit, dependencyVersion }) => {
 		const ext = typescript ? 'ts' : 'js';
 
-		sv.dependency('@oslojs/crypto', '^1.0.1');
-		sv.dependency('@oslojs/encoding', '^1.1.0');
+		sv.devDependency('@oslojs/crypto', '^1.0.1');
+		sv.devDependency('@oslojs/encoding', '^1.1.0');
 
 		if (options.demo) {
 			// password hashing for demo
@@ -142,11 +142,7 @@ export default defineAddon({
 			if (drizzleDialect === 'sqlite' || drizzleDialect === 'turso') {
 				js.imports.addNamed(ast, {
 					from: 'drizzle-orm/sqlite-core',
-					imports: {
-						sqliteTable: 'sqliteTable',
-						text: 'text',
-						integer: 'integer'
-					}
+					imports: ['sqliteTable', 'text', 'integer']
 				});
 				js.object.overrideProperties(userAttributes, {
 					properties: {
@@ -176,11 +172,7 @@ export default defineAddon({
 			if (drizzleDialect === 'mysql') {
 				js.imports.addNamed(ast, {
 					from: 'drizzle-orm/mysql-core',
-					imports: {
-						mysqlTable: 'mysqlTable',
-						varchar: 'varchar',
-						datetime: 'datetime'
-					}
+					imports: ['mysqlTable', 'varchar', 'datetime']
 				});
 				js.object.overrideProperties(userAttributes, {
 					properties: {
@@ -212,11 +204,7 @@ export default defineAddon({
 			if (drizzleDialect === 'postgresql') {
 				js.imports.addNamed(ast, {
 					from: 'drizzle-orm/pg-core',
-					imports: {
-						pgTable: 'pgTable',
-						text: 'text',
-						timestamp: 'timestamp'
-					}
+					imports: ['pgTable', 'text', 'timestamp']
 				});
 				js.object.overrideProperties(userAttributes, {
 					properties: {
@@ -260,20 +248,17 @@ export default defineAddon({
 			const { ast, generateCode } = parseScript(content);
 
 			js.imports.addNamespace(ast, { from: '$lib/server/db/schema', as: 'table' });
-			js.imports.addNamed(ast, { from: '$lib/server/db', imports: { db: 'db' } });
+			js.imports.addNamed(ast, { from: '$lib/server/db', imports: ['db'] });
 			js.imports.addNamed(ast, {
 				from: '@oslojs/encoding',
-				imports: {
-					encodeBase64url: 'encodeBase64url',
-					encodeHexLowerCase: 'encodeHexLowerCase'
-				}
+				imports: ['encodeBase64url', 'encodeHexLowerCase']
 			});
-			js.imports.addNamed(ast, { from: '@oslojs/crypto/sha2', imports: { sha256: 'sha256' } });
-			js.imports.addNamed(ast, { from: 'drizzle-orm', imports: { eq: 'eq' } });
+			js.imports.addNamed(ast, { from: '@oslojs/crypto/sha2', imports: ['sha256'] });
+			js.imports.addNamed(ast, { from: 'drizzle-orm', imports: ['eq'] });
 			if (typescript) {
 				js.imports.addNamed(ast, {
 					from: '@sveltejs/kit',
-					imports: { RequestEvent: 'RequestEvent' },
+					imports: ['RequestEvent'],
 					isType: true
 				});
 			}