+++ packages/addons/storybook/index.ts
@@ -12,7 +12,7 @@ export default defineAddon({
 		runsAfter('eslint');
 	},
 	run: async ({ sv }) => {
-		const args = ['storybook@latest', 'init', '--skip-install', '--no-dev'];
+		const args = ['create-storybook@latest', '--skip-install', '--no-dev'];
 
 		// skips the onboarding prompt during tests
 		if (process.env.NODE_ENV?.toLowerCase() === 'test') args.push('--yes');