+++ packages/cli/commands/add/utils.ts
@@ -111,6 +111,7 @@ export const commonFilePaths = {
 	packageJson: 'package.json',
 	svelteConfig: 'svelte.config.js',
 	tsconfig: 'tsconfig.json',
+	viteConfig: 'vite.config.js',
 	viteConfigTS: 'vite.config.ts'
 } as const;