+++ packages/cli/utils/common.ts
@@ -4,6 +4,7 @@ import * as p from '@clack/prompts';
 import type { Argument, HelpConfiguration, Option } from 'commander';
 import { UnsupportedError } from './errors.ts';
 import process from 'node:process';
+import { isVersionUnsupportedBelow } from '@sveltejs/cli-core';
 
 const NO_PREFIX = '--no-';
 let options: readonly Option[] = [];
@@ -75,6 +76,15 @@ type MaybePromise = () => Promise<void> | void;
 export async function runCommand(action: MaybePromise): Promise<void> {
 	try {
 		p.intro(`Welcome to the Svelte CLI! ${pc.gray(`(v${pkg.version})`)}`);
+
+		const minimumVersion = '18.3.0';
+		const unsupported = isVersionUnsupportedBelow(process.versions.node, minimumVersion);
+		if (unsupported) {
+			p.log.warn(
+				`You are using Node.js ${pc.red(process.versions.node)}, please upgrade to Node.js ${pc.green(minimumVersion)} or higher.`
+			);
+		}
+
 		await action();
 		p.outro("You're all set!");
 	} catch (e) {
@@ -104,4 +114,24 @@ export function forwardExitCode(error: unknown) {
 	} else {
 		process.exit(1);
 	}
+}
+
+export function parseAddonOptions(optionFlags: string | undefined): string[] | undefined {
+	// occurs when an `=` isn't present (e.g. `sv add foo`)
+	if (optionFlags === undefined || optionFlags === '') {
+		return undefined;
+	}
+
+	// Split on + and validate each option individually
+	const options = optionFlags.split('+');
+
+	// Validate that each individual option follows the name:value pattern
+	const malformed = options.filter((option) => !/.+:.*/.test(option));
+
+	if (malformed.length > 0) {
+		const message = `Malformed arguments: The following add-on options: ${malformed.map((o) => `'${o}'`).join(', ')} are missing their option name or value (e.g. 'addon=option1:value1+option2:value2').`;
+		throw new Error(message);
+	}
+
+	return options;
 }