+++ packages/core/addon/config.ts
@@ -20,7 +20,7 @@ export type Scripts<Args extends OptionDefinition> = {
 };
 
 export type SvApi = {
-	pnpmBuildDependendency: (pkg: string) => void;
+	pnpmBuildDependency: (pkg: string) => void;
 	dependency: (pkg: string, version: string) => void;
 	devDependency: (pkg: string, version: string) => void;
 	execute: (args: string[], stdio: 'inherit' | 'pipe') => Promise<void>;
@@ -78,13 +78,40 @@ export type TestDefinition<Args extends OptionDefinition> = {
 	condition?: (options: OptionValues<Args>) => boolean;
 };
 
-export function defineAddonOptions<const Args extends OptionDefinition>(options: Args): Args {
-	return options;
-}
-
 type MaybePromise<T> = Promise<T> | T;
 
-export type Precondition = {
+export type Verification = {
 	name: string;
 	run: () => MaybePromise<{ success: boolean; message: string | undefined }>;
 };
+
+type Prettify<T> = {
+	[K in keyof T]: T[K];
+} & unknown;
+
+// Builder pattern for addon options
+export type OptionBuilder<T extends OptionDefinition> = {
+	add<K extends string, const Q extends Question<T & Record<K, Q>>>(
+		key: K,
+		question: Q
+	): OptionBuilder<T & Record<K, Q>>;
+	build(): Prettify<T>;
+};
+
+// Initializing with an empty object is intended given that the starting state _is_ empty.
+// eslint-disable-next-line @typescript-eslint/no-empty-object-type
+export function defineAddonOptions(): OptionBuilder<{}> {
+	return createOptionBuilder({});
+}
+
+function createOptionBuilder<const T extends OptionDefinition>(options: T): OptionBuilder<T> {
+	return {
+		add(key, question) {
+			const newOptions = { ...options, [key]: question };
+			return createOptionBuilder(newOptions);
+		},
+		build() {
+			return options;
+		}
+	};
+}