+++ packages/core/tests/js/imports/named-import-array/output.ts
@@ -1,0 +1,2 @@
+import { Handle } from '@sveltejs/kit';
+import { namedOne } from 'package';