+++ packages/core/tests/js/imports/named-import/output.ts
@@ -1,2 +1,3 @@
+import { foo as bar, baz } from 'xyz';
 import { Handle } from '@sveltejs/kit';
 import { namedOne } from 'package';