+++ packages/core/tests/js/imports/named-import/run.ts
@@ -6,4 +6,6 @@ export function run(ast: AstTypes.Program): void {
 	imports.addNamed(ast, { from: '@sveltejs/kit', imports: { Handle: 'Handle' }, isType: false });
 	// adding the same import twice should not produce two imports
 	imports.addNamed(ast, { from: '@sveltejs/kit', imports: { Handle: 'Handle' }, isType: false });
+
+	imports.addNamed(ast, { from: 'xyz', imports: { foo: 'bar', baz: 'baz' } });
 }