+++ packages/core/tests/js/vite/add-plugin-mode/output.ts
@@ -1,0 +1,15 @@
+import { firstPlugin } from 'first-plugin';
+import lastPlugin from 'last-plugin';
+import middlePlugin from 'middle-plugin';
+import { sveltekit } from '@sveltejs/kit/vite';
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+	plugins: [
+		firstPlugin(),
+		// a default plugin
+		sveltekit(),
+		middlePlugin(),
+		lastPlugin()
+	]
+});