+++ packages/core/tests/js/vite/function-return/output.ts
@@ -1,0 +1,8 @@
+import myPlugin from 'my-plugin';
+import { sveltekit } from '@sveltejs/kit/vite';
+import { defineConfig } from 'vite';
+
+export default defineConfig(() => {
+	// Some stuff here
+	return { plugins: [sveltekit(), myPlugin()] };
+});